add new playbook for rogue wireless
This playbook helps the security team detect and remove rogue wireless
access points (WAPs). A Raspberry Pi 3 with a battery pack is used as
a mobile scanner, periodically polling the 2.4 GHz frequency and
collecting the ESSID (network name), MAC address, channel, signal
strength, signal quality, and security protocol of each WAP in range.
The results of each scan are compared against a whitelist of
known-good network names and a greylist of potential "evil twin"
network names. The results are tracked live using artifacts and
Heads-up Display pins posted to the most recently created Case with
the label "wireless". Finally, a team member is tasked to carry around
the Raspberry Pi and use the live feed of MAC addresses and signal
strengths to locate the rogue WAPs and remove them.