add new McAfee phishing investigate playbook
First written as an example playbook for BlackHat 2017, this
playbook investigates a potential phishing email with a suspicious
attachment. Execution starts when a suspicious email with an attachment
is forwarded to an inbox which Phantom is polling. McAfee Advanced
Threat Defense is used to do static and dynamic analysis on any
attachments to the email. If the attachment(s) appear malicious, any IP
addresses identified during the ATD detonation are used to do
geolocation and IP address reputation, further enriching the data in the
original container. The ATD detonation also records the hashes of files
that are created or touched. McAfee Active Response is used to search
within the enterprise for those hashes, and if there are any matches
they are shared on the OpenDXL message fabric.

Originally developed by Martin Ohl, Solution Architect EMEA at McAfee