add Website Unblock Request playbook
This playbook is designed to facilitate a Website Unblock type of
request from a ticketing system.  It is designed to operate on
containers that are ingested from a ticketing platform such as JIRA or
ServiceNow that contain the website URL requested as well as the
requesting user.  The playbook performs some 3rd party reputation
lookups as well as a BlueCoat WebPulse categorization request.
Appropriate notifications are sent to users and the original request
ticket is updated at the end based on the outcome.