add MacOS Root Password Mitigate playbook
Responds to CVE-2017-13872 by using SSH to access MacOS endpoints.
First the following information is gathered: MacOS version number,
whether TCP port 5900 is open, whether the root account is enabled,
and when the root account password was last changed. The gathered
information is then compiled into an email and Zendesk ticket for
analysis. Then through a prompt an analyst can decide whether to
disable the ARD remote screen sharing service and/or add a long,
random password for the root account.