add VMWorld Wannacry Response playbook
In virtualized environments the ability to take snapshots of VM's
allows incident response systems to save crucial data from encryption,
but only if the snapshot is triggered before encryption finishes. This
is an area where security automation can play a particularly important
role. In this playbook Phantom receives an alert from Vectra Cognito
signifying that a ransomware variant such as Wannacry has been
detected on one or more endpoints within the network. Phantom
immediately snapshots the affected devices that are virtual machines
managed by vSphere, and blocks command and control traffic using the
VMWare NSX virtual firewall. Beyond that immediate response the
security team is also asked whether or not to quarantine the affected
hosts with Carbon Black.