add VMWorld C2 Response playbook
This Playbook demonstrates a response to a command and control alert
when an organization is operating a VMWare virtualized environment.
The individual endpoints are vSphere guests and the network stack is
virtualized with VMWare NSX. This allows unique response options
including virtual machine snapshots and the flexible deployment of new
firewall rules across dynamic infrastructure. We built this Playbook
in a partnership with Rackspace and VMWare. As one of the largest
vSphere service providers, Rackspace provides experienced insight into
the deployment of security automation and orchestration in the
virtualized datacenter.