add Botnet Report Monitor playbook
This Playbook connects to multiple PAN Firewalls to monitor for
behavior-based botnet detections. Any detections found by the PAN
Firewall are processed by this Playbook if they have a confidence
score higher than the configured value. First each of the detections
is checked against a whitelist of known false positives and if the
traffic source is not whitelisted the botnet detection is used to
create a ticket in ServiceNow and notify collaborators using Slack.

Author: Irek Romaniuk (minor changes by the Phantom team)