add Nagios Service Monitor playbook
This Playbook enables automatic response to notifications from a
Nagios server that is monitoring Linux services. When Nagios detects
service disruption it is configured to send an email to Phantom, which
will trigger this Playbook to attempt a quick fix for the issue by
using SSH to restart the System V service in question. The target
hostname/IP and service name are parsed from the subject line of the
email from Nagios then each parameter is checked against a whitelist
(a Custom List on Phantom) before sending the SSH command.

Community Contributor: Lora Fulton