add Vectra Advanced Block Host playbook
Process requests created by Vectra to either block or unblock the
specified IP address on a Palo Alto Networks Firewall. If the detected
device is a virtual machine and the request is to block the IP then a
snapshot will be taken and a further check will be done to see if the
virtual machine should be suspended.

Author: Chris Johnson