add ExtraHop New DNS Servers playbook
This playbook gets initiated by a Timer asset (setup for the Timer
asset is in the Notes section below) every 30 minutes to query
ExtraHop for any new DNS servers that have been seen on your network
in the last 30 minutes.  If one or more new DNS servers are discovered
then the playbook will automatically initiate a Nessus endpoint scan
on the IP of the new DNS server.  In addition the playbook will
retrieve the following information about the new dns server:
  - ExtraHop device object
  - List of peer devices communicated with in the last 30 minutes
  - List of client and server protocols spoken in the last 30 minutes