add ExtraHop Detect Data Exfiltration
This playbook processes an ExtraHop Addy anomaly indicating potential
data exfiltration on the network.  It first retrieves all of the peers
acting as a client in the last 30 minutes for the device that
triggered the anomaly.  Then it filters out private IP Addresses as
defined in RFC1918.  Next it looks up IP reputation scores for each of
the non-private IP Addresses that have communicated with the device
that triggered the anomaly in the last 30 minutes. If a known-bad IP
is found then that device will be tagged with "bad_ip_reputation" in
ExtraHop and a Phantom task will be created to track further manual
investigation of this event.