add new onboarding demo playbook
Conduct enrichment on the URL, file hash, IP address, and/or domain
name in the event. Use the results to add context by updating the
heads-up display and changing the severity of the event. This Playbook
can be updated as needed to provide further enrichment, automate
collaboration by sending emails, chat messages, or service tickets, or
even to respond automatically to the event with a containment or
correction action.