add Zscaler Hunt and Block URL playbook
When a suspicious URL is detected, this playbook can be used to
identify internal devices that have accessed that URL and triage the
organizational importance of those devices. Then, depending on the
maliciousness of the URL and whether or not the affected device
belongs to an executive in the organization,  the URL will be blocked
and an appropriate ServiceNow ticket will be created.