add let's encrypt domain investigate
Investigate domain names and URLs of a potentially malicious website.
These domain names and URLs could come from anywhere, but this
Playbook was designed to work with the Splunk Analytic Story called
"Common Phishing Frameworks", which focuses on evilginx2 phishing
techniques that harvest credentials from fake login sites. The full
investigation is only completed if at least one of the TLS
certificates of the domains matches the issuer distinguished name of
Let's Encrypt, which is a free service that provides automatically
issued TLS certificates. This Playbook gathers certificate information
for the domains, queries whois for the domains, takes a screenshot of
each of the URLs, and does a urlscan.io scan of each of the URLs.
Finally, all the results are formatted together and posted to the
event comments.