add EC2 Instance Investigation playbook
Investigate an AWS Security Hub finding related to an exposed EC2
instance which is being probed by potentially malicious traffic.
Gather information about the EC2 configuration, the activity on the
server, and any remote IP addresses that are directing traffic at the
server. Notify and assign the appropriate people using a Jira ticket
and a Slack message, then initiate a prompt to ask a responder whether
or not the EC2 instance should be moved to an isolated EC2 Security
Group using another playbook called "EC2 Instance Isolation".