add Recorded Future threat hunting
Starting with a single IP address, this playbook gathers a list of
related IP addresses, domain names, file hashes, and vulnerability
CVE's from Recorded Future. Then Splunk is used to build threat
hunting lookup tables and search across multiple data sources for
events containing the related entities. Finally, IP addresses are
blocked if approved by an analyst and an email is sent to notify a
responder if more than 10 of a certain kind of entity are matched at
once.