new playbook GCP Unusual Serviceaccount Usage
Investigate and respond to an unusual service account usage alert in
Google Cloud Platform. Gather information about the service account
keys associated with the service account in question, as well as the
compute instance involved in the activity, if applicable. According to
a prompt response from an analyst, optionally delete the service
account keys and/or stop the compute instance.