new playbook endace_splunk_search_download_pcap
This playbook demonstrates the integration between Splunk SOAR and
EndaceProbe. It operates on syslog messages forwarded by the Phantom
App for Splunk. The playbook queries EndaceProbe for packets of
interests and on successful match downloads PCAP files in the Splunk
SOAR local storage. The downloading of actual files is conditional and
depends upon event severity and/or number of Bytes matched (default is
10MB and can be changed). For PCAPs greater than 10MB, a EndaceP2P
download link is provided as a 'note' to the user.

Thanks to Soumen Chakraborty for this contribution!