Update datetime_modify.py to use utcnow
Update datetime_modify.py to use utcnow if the user doesn't provide a value for input_datetime. The current logic doesn't work if the system timezone has been modified from UTC.

Currently, if input_datetime and input_format_string aren't provided, this function uses now() to get the local time and a UTC strftime format to record that local time with a UTC 'Z' timezone value. The only way to use datetime_modify to get the current time - or do maths against it - is if you provided an input_format_string with the time offset specified, which isn't practical when you have to factor in DST.