New custom function - list zip
New custom function called "list_zip":

"This function zips two or more lists together to create a list of equal length. This can be useful when multiple upstream blocks are used for a single downstream block with multiple inputs. A maximum of 9 lists can be zipped together. The input lists are intended to be flat lists of strings, not nested lists or dictionaries."